requirements.txt
gradio>=3.0
pysentimiento
pillow
numpy
trasnformers
torch
pillow